+++ src/controller/login.ts
@@ -1,21 +1,23 @@
 import { BaseContext } from 'koa';
 import { Responses } from './../types';
 
-import UserModel from './../model/user';
+import UserModel, {UserDocument, UserType} from './../model/user';
+
+type InputBodyType = {email: string, password: string};
 
 class Login {
   public static create = async (ctx: BaseContext) => {
     const { 
-      email, password 
-    } = ctx.request.body;
+      email, password
+    } = <InputBodyType>ctx.request.body;
 
-    const user = await UserModel.findOne({email: email});
+    const user:UserDocument|null = await UserModel.findOne({email: email});
 
     if (user) {
       const isMatched = await user.comparePassword(password).catch(err => null);
       if (isMatched === true) {
-        const {id}  = user.toNormalization();
-        const token = await ctx.jwt.sign({id: id});
+        const {id}         = <UserType>user.toNormalization();
+        const token:string = await ctx.jwt.sign({id: id});
 
         return ctx.respond(200, { token, user_id: id });
       } else {