+++ CHANGELOG.md
@@ -4,6 +4,19 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## 1.1.0 - 2019-05-25
+
+### Added
+ - Added more types for the User Model
+ - Added health endpoint
+ - Added User update route
+### Changed
+ - N/A
+
+### Removed
+ - N/A
+
+
 ## 1.0.0 - 2019-05-24
 
 ### Added